requirements.txt
altair==4.2.0
click==8.0.3
matplotlib==3.5.1
numpy==1.21.5
pandas==1.3.5
Pillow==9.2.0
PyYAML==6.0
scipy==1.1.0
seaborn==0.11.2
six==1.16.0
tensorboardX==2.5.1
torch==1.10.1
torchvision==0.11.2
tqdm==4.62.3
